requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
dramatiq[redis,watch]==1.17.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pydantic-settings==2.7.0
httpx==0.28.1
redis==5.2.1
openai==1.59.3
anthropic==0.42.0
google-generativeai==0.8.3
python-json-logger==3.2.1
anyio==4.7.0
